Fix numpy 2.x mypy incompatibility, validated on the declared Python 3.12
pyproject.toml has declared requires-python = ">=3.12" all along, but
this dev environment only had 3.10.8 installed (worked around all
session via a tomllib backport shim, since tomllib is 3.11+ stdlib).
Installed a real Python 3.12.10 and created a project .venv against it
as the first step of TASK-097 (core service freezing) -- PyInstaller
bundles whatever interpreter builds it, so packaging needs to target
the actual declared runtime, not the workaround.

Reinstalling into a clean 3.12 environment pulled in current dependency
versions (numpy 2.5, onnxruntime 1.28, mypy 2.3, etc.) rather than
whatever had been resolved against 3.10 previously. Full suite (pytest,
ruff, import-linter) passed unchanged; mypy --strict surfaced one real
issue: newer numpy's overload resolution for `ndarray - ndarray` in
clip_embedding_provider.py/caption_provider.py's normalization step
infers `complex128` rather than the actual runtime `float32`, since the
intermediate expression's dtype isn't pinned. Fixed by adding an
explicit `.astype(np.float32)` on the normalization result (a no-op at
runtime, since the values are already float32) -- verified against the
real-model integration tests, not just the type-checker.

+++ src/core/infrastructure/caption_provider.py
@@ -156,6 +156,8 @@ def _preprocess_image(path: Path) -> NDArray[np.float32]:
 
     resized = rgb.resize((IMAGE_SIZE, IMAGE_SIZE), Image.Resampling.BILINEAR)
     array = np.asarray(resized, dtype=np.float32) / np.float32(255.0)
-    array = (array - np.array(IMAGE_MEAN, dtype=np.float32)) / np.array(IMAGE_STD, dtype=np.float32)
+    array = (
+        (array - np.array(IMAGE_MEAN, dtype=np.float32)) / np.array(IMAGE_STD, dtype=np.float32)
+    ).astype(np.float32)
     array = array.transpose(2, 0, 1).astype(np.float32)
     return array[np.newaxis, :, :, :]
